Documentation updates and bug fixes
Former-commit-id: f6dd786727032cde8b9e224e9762a258e3ef7f71 [formerly 2d65ad32e90fc67700e84c65bb2f67fad052902e]
Former-commit-id: 0bd9a197ebe131a4434e1a585f50066aaa452718

+++ src/librapid/array/stride.cpp
@@ -23,7 +23,7 @@ namespace librapid {
 		// Initialize members
 		m_isTrivial	   = true;
 		m_isContiguous = true;
-		m_dims		   = data.size();
+		m_dims		   = (int64_t)data.size();
 
 		// Check for a valid number of dimensions
 		if (m_dims > LIBRAPID_MAX_DIMS)
